books/go-pratice-introduction/chapter-8-a/main.go
package main

import (
	"encoding/json"
	"fmt"
	"io"
	"os"
)

type Post struct {
	Id       int       `json:"id"`
	Content  string    `json:"content"`
	Author   Author    `json:"author"`
	Comments []Comment `json:"comments"`
}

type Author struct {
	Id   int    `json:"id"`
	Name string `json:"name"`
}

type Comment struct {
	Id      int    `json:"id"`
	Content string `json:"content"`
	Author  string `json:"author"`
}

func decode(filename string) (post Post, err error) {
	jsonFile, err := os.Open(filename)
	if err != nil {
		return Post{}, err
	}
	defer jsonFile.Close()

	decoder := json.NewDecoder(jsonFile)
	for {
		err = decoder.Decode(&post)
		if err == io.EOF {
			break
		}
		if err != nil {
			fmt.Println("Error decoding JSON:", err)
			return Post{}, err
		}
	}
	return post, nil
}

func main() {
	_, err := decode("./post.json")
	if err != nil {
		fmt.Println("Error:", err)
	}
}
